refactor(markdown-docx): merge methods for roundtripping conversion between OOXML and CiceroMark into one class - `OOXMLTransformer` (#425)
